Move image processing from NetworkClient to Pin
NetworkClient should only deal with fetching data from the remote pinry
service and handing it over to the SyncAdapter (moving image around is
an exception for now). It shouldn't process image and generate its
thumbnail. A better place to do that would be a Pin class, at least fow
now.