Remove superfluous Cursor.moveToFirst call
Don't call moveToFirst, as we don't have to actually pull data from the
Database, we are only interested in the number of rows that the query
returned.

Also, try to make this piece of code a little easier on the eyes by
removing else clause, and moving Cursor.close() outside of it.